show the owner in repo_tree_view
